- Move `responder` implementations to `batcher.py` - Move `responder` instance from `LlmTask` to `LlmInvocationProcess`
